+++ src/lib.rs
@@ -208,21 +208,23 @@ pub const CODEC_ID_STR: &str = "pcx";
 pub use dcx::{encode_dcx, parse_dcx, DcxImage, DCX_MAGIC, DCX_MAX_PAGES};
 pub use decoder::{
     parse_pcx, parse_pcx_indexed_1bpp_3planes, parse_pcx_indexed_1bpp_4planes,
-    parse_pcx_indexed_2bpp_cga, parse_pcx_indexed_4bpp, parse_pcx_indexed_8bpp,
+    parse_pcx_indexed_2bpp_cga, parse_pcx_indexed_2bpp_cga_cpi, parse_pcx_indexed_4bpp,
+    parse_pcx_indexed_8bpp,
 };
 pub use encoder::{
     encode_pcx_1bpp_3planes_ega_rgb, encode_pcx_1bpp_4planes_ega, encode_pcx_1bpp_mono,
     encode_pcx_1bpp_mono_dpi, encode_pcx_24bpp, encode_pcx_24bpp_dpi, encode_pcx_24bpp_image,
     encode_pcx_24bpp_screen, encode_pcx_24bpp_window, encode_pcx_24bpp_window_dpi,
-    encode_pcx_24bpp_window_dpi_screen, encode_pcx_2bpp_cga, encode_pcx_4bpp_packed,
-    encode_pcx_8bpp_grayscale, encode_pcx_8bpp_grayscale_dpi, encode_pcx_8bpp_indexed,
-    encode_pcx_8bpp_indexed_dpi,
+    encode_pcx_24bpp_window_dpi_screen, encode_pcx_2bpp_cga, encode_pcx_2bpp_cga_cpi,
+    encode_pcx_4bpp_packed, encode_pcx_8bpp_grayscale, encode_pcx_8bpp_grayscale_dpi,
+    encode_pcx_8bpp_indexed, encode_pcx_8bpp_indexed_dpi,
 };
 pub use error::{PcxError, Result};
 pub use image::{
-    Pcx1bpp3PlanesPaletteSource, Pcx1bpp4PlanesPaletteSource, Pcx2bppCgaPaletteSource,
-    Pcx4bppPaletteSource, PcxImage, PcxIndexed1x3, PcxIndexed1x4, PcxIndexed2x1Cga, PcxIndexed4,
-    PcxIndexed8, PcxPaletteSource, PcxPixelFormat,
+    Pcx1bpp3PlanesPaletteSource, Pcx1bpp4PlanesPaletteSource, Pcx2bppCgaCpi,
+    Pcx2bppCgaPaletteSource, Pcx4bppPaletteSource, PcxImage, PcxIndexed1x3, PcxIndexed1x4,
+    PcxIndexed2x1Cga, PcxIndexed2x1CgaCpi, PcxIndexed4, PcxIndexed8, PcxPaletteSource,
+    PcxPixelFormat,
 };
 pub use types::{
     find_vga_palette, parse_header, PcxHeader, PCX_ENCODING_RLE, PCX_HEADER_SIZE, PCX_MANUFACTURER,